pyproject.toml
[project]
name = "rivals-tournament-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautiful-date>=2.3.0",
    "gcsa>=2.5.1",
    "python-dotenv>=1.0.1",
]
